Implement Auth using policies

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,9 @@ use Illuminate\Http\Request;
 
 class QuestionsController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -68,6 +71,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize("update", $question);
         return view('questions.edit', compact('question'));
     }
 
@@ -80,6 +84,7 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question);
         $question->update($request->only('title', 'body'));
         return redirect(route('questions.index'))->with('success', "Your question has been updated");
     }
@@ -92,7 +97,8 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-     $question->delete();
-     return redirect(route('questions.index'))->with('success', "Your question has been deleted");
+        $this->authorize("delete", $question);
+        $question->delete();
+        return redirect(route('questions.index'))->with('success', "Your question has been deleted");
     }
 }
